requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
django-redis==5.3.0
django-redisboard==8.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-metadata==6.8.0
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
